Move `InferenceConfig` to `models`
